SystemUI: Fix BT icon duplicate issue
The default code will add new icon to the beginning of the icon list,
this will change the index of the existed icon which will make the
system think it to be new and thus resulting duplicate icon issue.

Change to add the new icon to the end of the list.

Change-Id: I49339f1e9c0a42b1af358c2f5fb7592e5cf9ce63
CRs-Fixed: 1046294